Moved pages into folders.
