+++ src/pages/index.ts
@@ -1,6 +1,6 @@
-export * from "./Articles";
-export * from "./Contact";
-export * from "./Exams";
-export * from "./Home";
+export * from "./articles";
+export * from "./contact";
+export * from "./exams";
+export * from "./home";
 export * from "./NotFound";
-export * from "./Resources";
+export * from "./resources";
