+++ src/pages/resources/index.tsx
